requirements.txt
Django==4.2.13
mysqlclient
pillow
djangorestframework
django_q
setuptools